- fixed compile error in convertor system
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@5601 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/packages/converter/smr2rdox/branches/dev_linux/update/update.cpp
@@ -204,7 +204,7 @@ UpdateMove::UpdateMove(CREF(Position) posFromBegin, CREF(Position) posFromEnd, C
 	, m_posTo       (posTo       )
 	, m_fileFrom    (fileFrom    )
 {
-	if (m_fileFrom == rdoModelObjectsConvertor::UNDEFINED_OUT)
+	if (static_cast<int>(m_fileFrom) == static_cast<int>(rdoModelObjectsConvertor::UNDEFINED_OUT))
 	{
 		m_fileFrom = getCurrentType();
 	}
